openbsd-6.9: do not install bash
`bash` was necessary when the idea was to copy the `run` scripts into the VM and call `bash` on those scripts from within the VM. This turned out to be difficult. For the moment, without a QEMU guest agent, copying files into the VM may be a challenge.

Furthermore, I wonder if running `bash` scripts inside OpenBSD might not be "the OpenBSD way" as it uses `ksh`.

I've decided that it's better to not install `bash` and, instead, default to OpenBSD's default shell.