Remove development section in README.md
I've documented this in the wiki, which I feel is a better choice for editing. This way, I can keep the git history a little cleaner from grammar and wording mistakes.

Plus, it makes the project README a little smaller and focused, which it 'ought to be. No need to burden users with the intricacies of development.