Apply resources to Vagrantfile, not Packer file.
It turns out I've been doing it all wrong: I should have been defining the CPU and RAM settings in the Vagrantfile, as that is where the VM guest will get its resource definitions from. The settings that I've been configuring in the Packer file are, confusingly, for _building_ the VM image, not _running_ it.