Move image-generating files to separate repo, mario-campos/emulate-images.
Most of the changes to the repo are to the image-generating files, so it makes sense to keep that isolated to a separate repository.

It also serves to separate the dual-purpose of the os directories, which were used both for the Action to find the Vagrantfile and also to contain the image-generating files.